Added interactivity to generated tiles
City generator takes in two references to doors and NPCs. When loading tiles , we can now choose to makes generated tiles an interactable by instantiating them. Each tile, for testing, generates a door at the location of the tile. For NPC's make sure to add the DialougeCanvasManager prefab and change the instatiation from door to npc